requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.11
redis==5.0.1
celery==5.3.4
kafka-python==2.0.2
pydantic>=2.0.0
pydantic-settings>=2.0,<3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
transformers==4.35.2
torch==2.10.0
detoxify==0.5.2
openai==1.3.7
whisper==1.1.10
opencv-python==4.8.1.78
librosa==0.10.1
python-socketio==5.16.1
tesseract==0.1.3
pillow==12.1.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.5.1
matplotlib
seaborn==0.13.0
aiofiles==23.2.1
httpx==0.25.2
prometheus-client==0.19.0
structlog==23.2.0
